requirements.txt
ga4gh-common
nose
